+++ sentry-android-replay/src/test/java/io/sentry/android/replay/viewhierarchy/ComposeMaskingOptionsTest.kt
@@ -15,8 +15,11 @@ import androidx.compose.material3.TextField
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.semantics.clearAndSetSemantics
+import androidx.compose.ui.semantics.editableText
 import androidx.compose.ui.semantics.invisibleToUser
 import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
@@ -52,6 +55,7 @@ class ComposeMaskingOptionsTest {
     fun setup() {
         System.setProperty("robolectric.areWindowsMarkedVisible", "true")
         ComposeMaskingOptionsActivity.textModifierApplier = null
+        ComposeMaskingOptionsActivity.textFieldModifierApplier = null
         ComposeMaskingOptionsActivity.containerModifierApplier = null
         ComposeMaskingOptionsActivity.fontSizeApplier = null
     }
@@ -84,6 +88,23 @@ class ComposeMaskingOptionsTest {
         val textNodes = activity.get().collectNodesOfType<TextViewHierarchyNode>(options)
         // the text should be laid out when fontSize is specified
         assertEquals("Random repo", (textNodes.first().layout as? ComposeTextLayout)?.layout?.layoutInput?.text?.text)
+    }
+
+    @Test
+    fun `when text input field is readOnly still masks it`() {
+        ComposeMaskingOptionsActivity.textFieldModifierApplier = {
+            // newer versions of compose basically do this when a TextField is readOnly
+            Modifier.clearAndSetSemantics { editableText = AnnotatedString("Placeholder") }
+        }
+        val activity = buildActivity(ComposeMaskingOptionsActivity::class.java).setup()
+        shadowOf(Looper.getMainLooper()).idle()
+
+        val options = SentryOptions().apply {
+            sessionReplay.maskAllText = true
+        }
+
+        val textNodes = activity.get().collectNodesOfType<TextViewHierarchyNode>(options)
+        assertTrue(textNodes[1].shouldMask)
     }
 
     @Test
@@ -220,6 +241,7 @@ private class ComposeMaskingOptionsActivity : ComponentActivity() {
 
     companion object {
         var textModifierApplier: (() -> Modifier)? = null
+        var textFieldModifierApplier: (() -> Modifier)? = null
         var containerModifierApplier: (() -> Modifier)? = null
         var fontSizeApplier: (() -> TextUnit)? = null
     }
@@ -243,6 +265,7 @@ private class ComposeMaskingOptionsActivity : ComponentActivity() {
                 )
                 Text("Random repo", fontSize = fontSizeApplier?.invoke() ?: TextUnit.Unspecified)
                 TextField(
+                    modifier = textFieldModifierApplier?.invoke() ?: Modifier,
                     value = TextFieldValue("Placeholder"),
                     onValueChange = { _ -> }
                 )